apps/ml-engine/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
prophet==1.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.32
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.1
apscheduler==3.10.4
requests==2.31.0
pytz==2024.1
